change db and add user

+++ P9/Models/Cuenta.cs
@@ -5,11 +5,11 @@ namespace P9.AutoModel
 {
   public partial class Cuenta
   {
-    public long Id { get; set; }
-    public long? NCuentaGerente { get; set; }
-    public long? NCuentaEmpleado { get; set; }
+    public int Id { get; set; }
+    public int? NCuentaGerente { get; set; }
+    public int? NCuentaEmpleado { get; set; }
     public long? NCuentaUsuario { get; set; }
-    public long Tipo { get; set; }
+    public int Tipo { get; set; }
 
     public virtual Empleado? NCuentaEmpleadoNavigation { get; set; }
     public virtual Gerente? NCuentaGerenteNavigation { get; set; }
